separate some methods from CommonUtils to FormatUtils
add CommonUtils back
