+++ examples/src/main/java/tachyon/examples/BasicOperations.java
@@ -4,9 +4,9 @@
  * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance with the License. You may obtain a
  * copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software distributed under the License
  * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
  * or implied. See the License for the specific language governing permissions and limitations under
@@ -33,6 +33,7 @@ import tachyon.client.TachyonFS;
 import tachyon.client.WriteType;
 import tachyon.conf.TachyonConf;
 import tachyon.util.CommonUtils;
+import tachyon.util.FormatUtils;
 
 public class BasicOperations implements Callable<Boolean> {
   private static final Logger LOG = LoggerFactory.getLogger(Constants.LOGGER_TYPE);
@@ -60,7 +61,7 @@ public class BasicOperations implements Callable<Boolean> {
     LOG.debug("Creating file...");
     long startTimeMs = CommonUtils.getCurrentMs();
     int fileId = tachyonClient.createFile(mFilePath);
-    CommonUtils.printTimeTakenMs(startTimeMs, LOG, "createFile with fileId " + fileId);
+    FormatUtils.printTimeTakenMs(startTimeMs, LOG, "createFile with fileId " + fileId);
   }
 
   private void writeFile(TachyonFS tachyonClient) throws IOException {
@@ -80,7 +81,7 @@ public class BasicOperations implements Callable<Boolean> {
     os.write(buf.array());
     os.close();
 
-    CommonUtils.printTimeTakenMs(startTimeMs, LOG, "writeFile to file " + mFilePath);
+    FormatUtils.printTimeTakenMs(startTimeMs, LOG, "writeFile to file " + mFilePath);
   }
 
   private boolean readFile(TachyonFS tachyonClient) throws IOException {
@@ -100,7 +101,7 @@ public class BasicOperations implements Callable<Boolean> {
     }
     buf.close();
 
-    CommonUtils.printTimeTakenMs(startTimeMs, LOG, "readFile file " + mFilePath);
+    FormatUtils.printTimeTakenMs(startTimeMs, LOG, "readFile file " + mFilePath);
     return pass;
   }
