+++ shell/src/test/java/tachyon/shell/TFsShellTest.java
@@ -44,6 +44,7 @@ import tachyon.conf.TachyonConf;
 import tachyon.master.LocalTachyonCluster;
 import tachyon.thrift.ClientBlockInfo;
 import tachyon.util.CommonUtils;
+import tachyon.util.FormatUtils;
 import tachyon.util.network.NetworkAddressUtils;
 
 /**
@@ -111,7 +112,7 @@ public class TFsShellTest {
     fos.write(toWrite);
     fos.close();
     mFsShell
-        .copyFromLocal(new String[] {"copyFromLocal", testFile.getAbsolutePath(), "/testFile"});
+        .copyFromLocal(new String[]{"copyFromLocal", testFile.getAbsolutePath(), "/testFile"});
     Assert.assertEquals(getCommandOutput(new String[] {"copyFromLocal", testFile.getAbsolutePath(),
         "/testFile"}), mOutput.toString());
     TachyonFile tFile = mTfs.getFile(new TachyonURI("/testFile"));
@@ -352,14 +353,14 @@ public class TFsShellTest {
     String expected = "";
     String format = "%-10s%-25s%-15s%-5s\n";
     expected +=
-        String.format(format, CommonUtils.getSizeFromBytes(10),
+        String.format(format, FormatUtils.getSizeFromBytes(10),
             TFsShell.convertMsToDate(files[0].getCreationTimeMs()), "In Memory",
             "/testRoot/testFileA");
     expected +=
-        String.format(format, CommonUtils.getSizeFromBytes(0),
+        String.format(format, FormatUtils.getSizeFromBytes(0),
             TFsShell.convertMsToDate(files[1].getCreationTimeMs()), "", "/testRoot/testDir");
     expected +=
-        String.format(format, CommonUtils.getSizeFromBytes(30),
+        String.format(format, FormatUtils.getSizeFromBytes(30),
             TFsShell.convertMsToDate(files[3].getCreationTimeMs()), "Not In Memory",
             "/testRoot/testFileC");
     Assert.assertEquals(expected, mOutput.toString());
@@ -381,14 +382,14 @@ public class TFsShellTest {
     String expected = "";
     String format = "%-10s%-25s%-15s%-5s\n";
     expected +=
-        String.format(format, CommonUtils.getSizeFromBytes(10),
+        String.format(format, FormatUtils.getSizeFromBytes(10),
             TFsShell.convertMsToDate(files[0].getCreationTimeMs()), "In Memory",
             "/testRoot/testFileA");
     expected +=
-        String.format(format, CommonUtils.getSizeFromBytes(0),
+        String.format(format, FormatUtils.getSizeFromBytes(0),
             TFsShell.convertMsToDate(files[1].getCreationTimeMs()), "", "/testRoot/testDir");
     expected +=
-        String.format(format, CommonUtils.getSizeFromBytes(30),
+        String.format(format, FormatUtils.getSizeFromBytes(30),
             TFsShell.convertMsToDate(files[2].getCreationTimeMs()), "Not In Memory",
             "/testRoot/testFileC");
     Assert.assertEquals(expected, mOutput.toString());
